requirements.txt
# Core dependencies for miDiscovery SAE training
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0
huggingface_hub>=0.17.0

# Data processing and numerical computing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# Visualization and analysis
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Progress tracking and utilities
tqdm>=4.66.0
wandb>=0.16.0

# Optional but recommended
jupyter>=1.0.0
ipywidgets>=8.0.0

# Development tools (optional)
black>=23.0.0
pytest>=7.4.0
mypy>=1.6.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0  # For database migrations
python-dotenv>=1.0.0  # For environment variables